move AKI, SKI
